+++ addon/components/searchbar-node.js
@@ -1,0 +1,9 @@
+import Component from "@ember/component";
+
+export default Component.extend({  
+  click(e) {
+    if (!e.target.href) {
+      e.stopPropagation();
+    }
+  }
+});